+++ packages/opencensus-exporter-jaeger/src/jaeger.ts
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-import {Exporter, ExporterBuffer, ExporterConfig, RootSpan, Span} from '@opencensus/core';
+import {Exporter, ExporterConfig, RootSpan, Span} from '@opencensus/core';
 import {logger, Logger} from '@opencensus/core';
 import * as os from 'os';
-
 import {spanToThrift, Tag, TagValue, ThriftProcess, ThriftUtils, UDPSender, Utils} from './jaeger-driver';
+
+const DEFAULT_BUFFER_FLUSH_INTERVAL_MILLIS = 1000;
+const DEFAULT_BUFFER_SIZE = 1000;
 
 /**
  * Options for Jaeger configuration
@@ -53,8 +55,6 @@ export class JaegerTraceExporter implements Exporter {
   /** Timeout control */
   /** Max time for a buffer can wait before being sent */
   private bufferTimeout: number;
-  /** Manage when the buffer timeout needs to be reseted */
-  private resetTimeout = false;
   /** Indicates when the buffer timeout is running */
   private timeoutSet = false;
 
@@ -62,8 +62,9 @@ export class JaegerTraceExporter implements Exporter {
   constructor(options: JaegerTraceExporterOptions) {
     const pjson = require('../../package.json');
     this.logger = options.logger || logger.logger();
-    this.bufferTimeout = options.bufferTimeout;
-    this.bufferSize = options.bufferSize;
+    this.bufferTimeout =
+        options.bufferTimeout || DEFAULT_BUFFER_FLUSH_INTERVAL_MILLIS;
+    this.bufferSize = options.bufferSize || DEFAULT_BUFFER_SIZE;
     this.sender = new UDPSender(options);
     const tags: Tag[] = options.tags || [];